speech to text/requirements.txt
accelerate==0.33.0
ai71==0.0.18
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.31.0
attrs==23.2.0
audio-recorder-streamlit==0.0.9
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorthief==0.2.1
cvlib==0.2.7
dataclasses-json==0.6.7
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
entrypoints==0.4
ffprobe==0.5
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.3.1
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib-metadata==6.11.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llama-cloud==0.0.9
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.56
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
openai==1.36.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
progressbar==2.5
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
PyMatting==1.1.12
pypdf==4.3.1
pyshortners==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
rembg==2.0.57
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
setuptools==70.2.0
shortzyy==0.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.37.0
streamlit-audiorecorder==0.0.5
streamlit-chat==0.1.1
streamlit-float==0.3.5
streamlit-lottie==0.0.5
striprtf==0.0.26
sympy==1.12.1
tenacity==8.4.2
tifffile==2024.7.2
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
transformers==4.43.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
utils==1.0.2
validators==0.33.0
virtualenv==20.26.3
watchdog==4.0.1
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
